Trail to double

+++ TrailAid.Data/Trail.cs
@@ -44,10 +44,10 @@ namespace TrailAid.Data
         }
         public string Difficulty { get; set; }
         public string Description { get; set; }
-        public int Distance { get; set; }
+        public double Distance { get; set; }
         public string TypeOfTerrain { get; set; }
         public string Tags { get; set; }
-        public int Elevation { get; set; }
+        public double Elevation { get; set; }
         public string RouteType { get; set; }
         public virtual List<Visited> Ratings { get; set; } = new List<Visited>();
     }
