+++ TrailAid.Models/Trail/TrailDetail.cs
@@ -20,10 +20,10 @@ namespace TrailAid.Models.Trail
         public double Rating { get; set; }
         public string Difficulty { get; set; }
         public string Description { get; set; }
-        public int Distance { get; set; }
+        public double Distance { get; set; }
         public string TypeOfTerrain { get; set; }
         public string Tags { get; set; }
-        public int Elevation { get; set; }
+        public double Elevation { get; set; }
         public string RouteType { get; set; }
     }
 }
